improvement of the code and rectification for the activity graph
