+++ src/Components/Layout/layout.module.scss
@@ -8,5 +8,5 @@
 .main {
   display: flex;
   width: 100%;
-  height: auto;
+  height: 100%;
 }
